Update Slot Machine (1).ipynb
This code mimicks the workings of a casino slot machine. The machine will ask how much you would like to bet, the lines you would like to bet on and how much you would like to bet on each line.
The code has been designed so that you cannot offer an element that is not a number and you cannot input a negative number. 
After you spin you will also be told your winnings (if any) and your outstanding balance.